Move traveling logic to World class
